broke down classes a bit
