+++ Grevit.Revit.2019/Grevit.Revit.2019.csproj
@@ -106,6 +106,10 @@
       <Link>Utilities.cs</Link>
     </Compile>
     <Compile Include="ExtEventHandler.cs" />
+    <Compile Include="Grevit\GrevitBuildModel.cs" />
+    <Compile Include="Grevit\GrevitCommand.cs" />
+    <Compile Include="Grevit\GrevitFaceReference.cs" />
+    <Compile Include="Grevit\ParameterNames.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
